First step of loading budget data Refactored calculation of week range: We only need week start as string that can be parsed to localdate

+++ myExpenses/src/main/java/org/totschnig/myexpenses/viewmodel/data/BudgetProgress.kt
@@ -4,8 +4,8 @@ data class BudgetProgress(
     val title: String,
     val allocated: Long,
     val spent: Long,
-    val totalDays: Int,
-    val currentDay: Int
+    val totalDays: Long,
+    val currentDay: Long
 ) {
     val remainingBudget: Long
         get() = allocated - spent
